move some methods to utils.py, to prepare STRIPS learning refactor (#708)
* move some methods to utils.py, to prepare STRIPS learning refactor

* checks
