+++ src/nsrt_learning/option_learning.py
@@ -7,13 +7,14 @@ import logging
 from typing import Dict, List, Sequence, Set, Tuple
 
 import numpy as np
+from gym.spaces import Box
 
 from predicators.src.envs import get_or_create_env
 from predicators.src.envs.blocks import BlocksEnv
 from predicators.src.ml_models import ImplicitMLPRegressor, MLPRegressor, \
     Regressor
 from predicators.src.settings import CFG
-from predicators.src.structs import Action, Array, Box, Datastore, Object, \
+from predicators.src.structs import Action, Array, Datastore, Object, \
     OptionSpec, ParameterizedOption, Segment, State, STRIPSOperator, \
     Variable
 from predicators.src.utils import OptionExecutionFailure
